==> 1 Data Collection/1 Scraping Birth Control App Reviews from the Apple App Store.ipynb <==
# 1 Scraping Birth Control App Data from the Apple App Store

- This notebook uses the app_store_scraper package to scrape the Apple app store for the reviews of 4 birth control apps to gain insight into women's current issues two years after the overturn of Roe vs. Wade.

- I uses the sklearn, pystemmer and eli5 packages to build a classifer to predict whether users who wrote 1-star and 2-star reviews on 4 apps expressed that they were [+ categories]

- Apps were selected based on search results for "birth control" on the app store, as well as online reviews of mental health apps. The dataframe containing their names and ids is manually created here.

## 1. Install packages to get prepared

#!pip install itunes-app-scraper-dmi==0.9.4
#!pip install app-store-scraper==0.3.5

import pandas as pd

from itunes_app_scraper.scraper import AppStoreScraper
from app_store_scraper import AppStore

import time
import re

import requests
from plotnine import * 

pd.set_option('display.max_columns', None)
pd.set_option("display.max_colwidth", 300)

## 2. Building dataframe of the 7 apps

This section builds a dataframe of the birth control apps to analyze from the App Store website (https://www.apple.com/app-store/). The names and ids of the apps are in each of their App store URLs

The apps are:
- APP 01: [Nurx: Birth Control Delivered](https://apps.apple.com/us/app/nurx-birth-control-delivered/id1213141301) 
- APP 02: [Natural Cycles: Birth Control](https://apps.apple.com/us/app/natural-cycles-birth-control/id765535549)
- APP 03: [myPill Birth Control Reminder](https://apps.apple.com/us/app/mypill-birth-control-reminder/id425632209)
- APP 04: [Birth Control Pill Reminder](https://apps.apple.com/us/app/birth-control-pill-reminder/id953740058) 
- APP 05: [Planned Parenthood Direct](https://apps.apple.com/us/app/planned-parenthood-direct/id1214393415)
- APP 06: [Flo Period & Pregnancy Tracker](https://apps.apple.com/us/app/flo-period-pregnancy-tracker/id1038369065)
- APP 07: [Clue Period Tracker & Calendar](https://apps.apple.com/us/app/clue-period-tracker-calendar/id657189652)


app_names_ids = [
    {'name':'nurx-birth-control-delivered',
    'id': 1213141301},
    {'name':'natural-cycles-birth-control',
    'id':765535549},
    {'name':'mypill-birth-control-reminder',
    'id': 425632209},
    {'name': 'birth-control-pill-reminder',
    'id':953740058},
    {'name': 'planned-parenthood-direct',
    'id':1214393415},
    {'name':'flo-period-pregnancy-tracker',
    'id': 1038369065},
    {'name':'clue-period-tracker-calendar',
    'id': 657189652},
]

birthControl_apps = pd.DataFrame(app_names_ids)
birthControl_apps

## 3. Scraping the reviews of 7 apps
- timeline: from 2020-05-02 to 2024-05-02 (4 years)
- number: all the reviews within the described timeline above

import datetime

all_7apps_reviews = []

for app in app_names_ids:
    my_app = AppStore(
      country='us',       
      app_name=app['name'], 
      app_id=app['id']    
    ) 

    my_app.review(
      after=pd.to_datetime('2020-05-02'),
      sleep=1,
    )
    
    reviews = my_app.reviews
    
    # Filter reviews to include only those before the end date
    end_date = datetime.datetime(2024, 5, 3)
    filtered_reviews = [review for review in reviews if review['date'] < end_date]
    
    for review in filtered_reviews:
        review['app_name'] = app['name']
        review['app_id'] = app['id']
        
    reviews_df = pd.DataFrame(filtered_reviews)
    
    all_7apps_reviews.append(reviews_df)
    time.sleep(5)


all_7apps_reviews_df = pd.concat(all_7apps_reviews)
all_7apps_reviews_df

#Quick Check1: How many reviews does each app have?
all_7apps_reviews_df.groupby('app_name').size().sort_values(ascending=False)

#Quick Check2: How many reviews are there for each rating level?
all_7apps_reviews_df.rating.value_counts()

#all_7apps_reviews_df.to_csv('all_7apps_reviews.csv')
