==> 1 Data Collection/3 Drag Unlabeled Data and Combine Labeled Data.ipynb <==
## 1. Drag Unlabeled Data

### 1) Group 1: Save the remaining data as "group1_unlabeled_data.csv"

import pandas as pd

# Load original dataset and training data
group1_df = pd.read_csv('group1_negative_reviews.csv') # all negative reviews from group 1
group1_training_data = pd.read_csv('group1_samples.csv') # selected sample reviews from group 1

# 'date' is the unique identifier for each review
group1_unlabeled_data = group1_df[~group1_df['date'].isin(group1_training_data['date'])]

# Save the unlabeled data to a CSV file
# group1_unlabeled_data.to_csv('group1_unlabeled_data.csv', index=False)

len(group1_unlabeled_data)

### 2) Group 2: Save the remaining data as "group2_unlabeled_data.csv"

# Load original dataset and training data
group2_df = pd.read_csv('group2_negative_reviews.csv') # all negative reviews from group 2
group2_training_data = pd.read_csv('group2_samples.csv') # selected sample reviews from group 2

# 'date' is the unique identifier for each review
group2_unlabeled_data = group2_df[~group2_df['date'].isin(group2_training_data['date'])]

# Save the unlabeled data to a CSV file
# group2_unlabeled_data.to_csv('group2_unlabeled_data.csv', index=False)

len(group2_unlabeled_data)

### 3) Group 1: Filter unlabeled data based on functinality

# Filter the data based on the 'app_name' column
group1_unlabeled_data_pt = group1_unlabeled_data[group1_unlabeled_data['app_name'].isin(['flo-period-pregnancy-tracker', 'clue-period-tracker-calendar'])]
group1_unlabeled_data_bc = group1_unlabeled_data[~group1_unlabeled_data['app_name'].isin(['flo-period-pregnancy-tracker', 'clue-period-tracker-calendar'])]

# Save the filtered data into separate CSV files
#group1_unlabeled_data_pt.to_csv('group1_collection2_pt_unlabeled.csv', index=False)
#group1_unlabeled_data_bc.to_csv('group1_collection1_bc_unlabeled.csv', index=False)

### 4) Group 2: Filter unlabeled data based on functinality

# Filter the data based on the 'app_name' column
group2_unlabeled_data_pt = group2_unlabeled_data[group2_unlabeled_data['app_name'].isin(['flo-period-pregnancy-tracker', 'clue-period-tracker-calendar'])]
group2_unlabeled_data_bc = group2_unlabeled_data[~group2_unlabeled_data['app_name'].isin(['flo-period-pregnancy-tracker', 'clue-period-tracker-calendar'])]

# Save the filtered data into separate CSV files
#group2_unlabeled_data_pt.to_csv('group2_collection2_pt_unlabeled.csv', index=False)
#group2_unlabeled_data_bc.to_csv('group2_collection1_bc_unlabeled.csv', index=False)

### 5)  Combine unlabeled datasets based on collection

# Concatenate the two dataframes
combined_pt_unlabeled = pd.concat([group1_unlabeled_data_pt, group2_unlabeled_data_pt], ignore_index=True)

# Save the combined dataframe to a new CSV file
#combined_pt_unlabeled.to_csv('combined_pt_unlabeled.csv', index=False)

# Concatenate the two dataframes
combined_bc_unlabeled = pd.concat([group1_unlabeled_data_bc, group2_unlabeled_data_bc], ignore_index=True)

# Save the combined dataframe to a new CSV file
#combined_bc_unlabeled.to_csv('combined_bc_unlabeled.csv', index=False)

## 2. Combine Labeled Data based on collection

# Load labeled data from Collection 1
group1_collection1_bc_tagged = pd.read_csv('group1_samples (beforeRoe) - group1_collection1_bc_samples.csv') 
group2_collection1_bc_tagged = pd.read_csv('group2_samples (postRoe) - group2_collection1_bc_samples.csv') 

# Concatenate the two dataframes
combined_bc_tagged = pd.concat([group1_collection1_bc_tagged, group2_collection1_bc_tagged], ignore_index=True)

# Save the combined dataframe to a new CSV file
#combined_bc_tagged.to_csv('combined_bc_tagged.csv', index=False)

# Load labeled data from Collection 2
group1_collection2_pt_tagged = pd.read_csv('group1_samples (beforeRoe) - group1_collection2_pt_samples.csv') 
group2_collection2_pt_tagged = pd.read_csv('group2_samples (postRoe) - group2_collection2_pt_samples.csv') 

# Concatenate the two dataframes
combined_pt_tagged = pd.concat([group1_collection2_pt_tagged, group2_collection2_pt_tagged], ignore_index=True)

# Save the combined dataframe to a new CSV file
#combined_pt_tagged.to_csv('combined_pt_tagged.csv', index=False)
